Integration of DigitalTwinState in MileStone flow and contract
The intended behavior is the following:
When creating an Arrival MileStone it is expected for all the related DT to be found as previously created states in the vault, and to be used as input states.
In other words:
- input: all the DT states related to a milestone;
- output: 1 milestone state and all the DT states with updated information;
This design choice has still to be confirmed.

MilestoneFlows.kt and MilestoneContract.kt have been edited to account for the aforementioned behavior.

In MilestoneFlowTests.kt two new basic tests to verify that the code runs have been written, but old tests still need to be updated according to the new behavior of the flow.