Redesign of Digital Twin State
Because of inheritance issues the first idea of reproducing accurately the same structure of the ontology within L0 has been abandoned.

The current structure is the following:
There is one state, the DT.
DT has "physicalObject" as attribute which defines the type of the DT (cargo, truck, etc.).
The DT implements all of the possible type (cargo, truck...) as data objects set to null by default.
Each data object (for instance Cargo) contains its specific attributes (for instance "natureOfCargo").
The contract will have to ensure correspondence between the type of the physical object (hence of the DT) and the data object that is initialized and thus different from null (yet to be implemented).

Files related to the old structure are removed.

When this line of the implementation will be self standing also the original DT will be replaced.