Discharge event now requires Load event as input
Flow and contract updated according to the new design implementation.

DT of type "cargo" in Load events have been constrained to 1, as we should avoid to run into a situation where we are discharging only a part of cargos from a Load event with multiple cargos.

Current tests have been updated accordingly.