Added skeleton of new Event structure
The new event structure is being built alongside the old one.
I find this way easier, in opposition of burning everything down and build from scratch.
Once this implementation is "complete" (in a basic way) I plan to remove the old one and refactor this one for a full substitution.