Including timestamps of START event when executing STOP event
When a STOP event is executed the flow looks for the
corresponding previous START event and provide it as
input, so it gets consumed. In order to make it easier
to find the relevant information of the whole history
of the association (from START to STOP) the timestamps
of the START event are copied in an ad-hoc field of
the STOP event.

State, flow and tests updated accordingly.