New timestamp model introduced, according to #64
Now the timestamp is no longer a simple linked hash but a more
complete object that includes ID, type and timestamp.
Identifiers of events are now linked to their timestamp,
while the UUID of an event state (containing all the timestamps of the
"same" event) is unlinked to the UUID (or ID) in RDF.
RDF event ID is, therefore, registered into the timestamp object.

- Parser adjusted accordingly
- Tests (contract, flow, service) adjusted accordingly
- Contract adjusted to account for more timestamps of the same type

To be completed:
Adjustment of flow to correctly address "update estimation",
"execution" or "creation of event" depending on the incoming event.