Changes to SetupServer readme
- Reorganized sections
I separated clearly the native-section from the docker-section. Default is native, but at the top of the document is mentioned that modules can be run also in dockers and to check out the related section at the bottom.

- Install JDK by default
To avoid confusion, I set as default in the examples the JDK instead of the JRE.

- Minor corrections to GDB instructions

- Set localhost as default for GDB (under the assumption that they run on the same machine and one would want to restrict access to it as much as possible)

- database.properties
 Instructions to change properties are now earlier in the readme and have been expanded a little. Later instructions recall that configuration back to ensure consistency

- More options on how to execute the node
Added the instructions to run a node without registering it to the NMS, and to run all the nodes on the same machine via de `runnodes` script

- Clarified where to find the jar of the client

- In the examples password is "test" as default